NNS1-2906: Remove unused transactions from debug store (#4770)
# Motivation

https://github.com/dfinity/nns-dapp/pull/4714 created a lot of dead
code.
If we try to clean if up all at once, it will be hard to make sure it's
really all deleted.
So I went over the removed imports from that PR and marked the ones that
aren't used anywhere else to be deleted.
And then only actually deleted dead code from the debug store.
The other dead code will be removed in future PRs because it will
generate further dead code and by making sure we always mark newly dead
code it will be easy to see that nothing is missed.

# Changes

1. Remove unused `debugTransactions` and related code.
2. Mark other dead code for deletion.

# Tests

Still pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary