NNS1-2966: Tooltip on voted neuron voting power (#4798)
# Motivation

Voting power of neurons is displayed as rounded numbers. We want to give
people access to the exact numbers.
This PR adds a tooltip on voting power of voted neurons, similar to what
https://github.com/dfinity/nns-dapp/pull/4713 did for votable neurons.

# Changes

1. Use `VotingPowerDisplay` component for the total voted neuron voting
power in `VotedNeuronList` and the individual voted neurons in
`MyVotes`.


# Tests

1. Unit tests for both tooltips were added in
`frontend/src/tests/lib/components/proposal-detail/VotingCard/VotedNeuronList.spec.ts`.
2. Page objects were updated.

# Todos

- [x] Add entry to changelog (if necessary).