NNS1-2905: Improve upgrade-downgrade-test (#4835)
# Motivation

We no longer store transactions in the nns-dapp canister because the
frontend now reads transactions from the index canister.
However we still store transaction indexes on individual accounts. These
indexes used to point into the vector of transactions which we no longer
have and are now completely unused. So we want to clean these up.
Because these accounts are stored in stable memory we have to be careful
about removing fields from them or we might get errors when decoding
stable memory.
In particular, we should be able to revert to a previous release of
nns-dapp in case of an emergency.
So we have to stop reading the fields at least 1 release before we stop
writing the fields and we can't just remove the fields directly.

However when testing the changes necessary for this, I noticed that our
`upgrade-downgrade-test` does not break even if we remove the fields
directly.

We used to encode all data pre-upgrade and decode all data post-upgrade,
but now accounts are stored in stable memory. This means that simply
upgrading is not enough to detect a data incompatibility. We have to
actually write accounts pre-upgrade and read them post-upgrade **and**
write accounts pre-downgrade and read accounts post-downgrade.

# Changes

1. In `scripts/nns-dapp/migration-test`, instead of populating all
accounts before upgrading and downgrading, populate half of the accounts
before upgrading and the other half between upgrading and downgrading.

# Tests

1. The `upgrade-downgrade-test` still passes on head of `main`.
2. The `upgrade-downgrade-test` now fails when removing transactions
fields from account, while previous it did not fail.

# Todos

- [x] Add entry to changelog (if necessary).