Fix description of snsdemo update PRs (#4841)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/4058 I removed a line from
the automatic PR description of snsdemo updates.
But I should have only removed part of that line.
We still check that the dfx version matches. Only the IC commit is moved
to a separate update job.

# Changes

Restore part of the line that was removed in
https://github.com/dfinity/nns-dapp/pull/4058

# Tests

no

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary