Support proposal type AddApiBoundaryNodesPayload  (#4854)
# Motivation

The proposal type to add API Boundary nodes [now
supports](https://gitlab.com/dfinity-lab/public/ic/-/merge_requests/19097)
adding multiple nodes in one proposal.
This is a breaking change but it's OK because proposals of this type
have never been created and will not be created for a few more weeks.

Also note that the candid in `rs/registry/canister/canister/canister.rs`
was regenerated automatically and because of how the tool works, this
resulted in `UpdateApiBoundaryNodesVersionPayload` being removed and
replaced by `AddApiBoundaryNodesPayload`, which is structurally the
same. We work around this in `rs/proposals/src/def.rs` with
```
pub type UpdateApiBoundaryNodesVersionPayload = crate::canisters::nns_registry::api::AddApiBoundaryNodesPayload;
```
Note that we were already in a similar situation with
`DeployGuestosToSomeApiBoundaryNodesPayload`.

# Changes

1. Ran `scripts/update_ic_commit --ic_commit
release-2024-05-09_23-02-storage-layer` to update candid files.
2. Ran `./scripts/proposals/did2rs` to update Rust types in `api.rs`
files.
3. Manually updated `def.rs` to rename the changed type and to work
around the type removed by the candid generation tool.
4. Rename the type in `rs/proposals/src/lib.rs`

# Tests

Added a test for `ic-admin propose-to-add-api-boundary-nodes`.

# Todos

- [x] Add entry to changelog (if necessary).