E2e test for following topics (#4867)
# Motivation

When adding a new topic in https://github.com/dfinity/nns-dapp/pull/4864
all tests passed but trying to add a followee for the new topic gave an
error before the governance canister was updated.
Things shouldn't be broken when all tests are passing so we need a test
for this.

# Changes

1. Add an e2e test which tries to add a followee for every existing
topic.
2. Add page objects.
3. Add test IDs.

# Tests

Tested it with a snapshot that does not have the new topic and ic-js
version that does have the new topic and saw the test fail.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary