Add NeuronsTableRow component (#4872)
# Motivation

Add more scaffolding for the neurons table without implementing anything
real.

# Changes

1. Add `TableNeuron` type (with just 1 field for now) to be used to
render neurons in the table. This abstracts away the difference between
NNS neurons and SNS neurons so we'll be able to reuse the table for SNS.
2. Add `tableNeuronsFromNeuronInfos` to convert NNS `NeuronInfo`s to
`TableNeuron`.
3. Create and pass `TableNeuron`s from `NnsNeurons` to `NeuronsTable`.
4. Add `NeuronsTableRow` component and use it in `NeuronsTable`
component.

# Tests

Added unit tests and page objects.

Drive-by: Fix `"should not render the neurons table"` test description
which was incorrectly copied.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet