Set canister IDs for app network (#4885)
# Motivation

We have a
[workflow](https://github.com/dfinity/nns-dapp/actions/workflows/deploy-to-app.yaml)
to deploy nns-dapp to a separate canister on mainnet which is not NNS
controlled, which we sometimes use for testing on mainnet.
It is currently not possible to use that workflow because some required
canisters don't have canister IDs specified for the `app` network.

# Changes

Specify canister IDs for `nns-index`, `cketh_ledger` and `cketh_index`
for the `app` network.

# Tests

Deployed to test ckUSDC.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary