Don't restore state while replica is still running. (#4887)
# Motivation

When running an snsdemo snapshot, your dfx state (including identities)
is replaced with state from the snapshot.
A `restore.sh` script is created, which is run when the replica is
stopped.
If something goes wrong the restore script can also be run manually.
But you shouldn't do this if the replica is still running.

# Changes

Add a check in the restore script whether the replica is still running.

# Tests

Ran a snapshot and tried to restore it manually while it was still
running.
Then checked that it got restored correctly when I stopped the replica.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary