Display 6 decimals for ckUSDC. (#4889)
# Motivation

Currently we always display tokens with 8 decimal precision on the
transaction review screen.
Even for ckETH, which supports 18 decimals because we don't think there
is value in displaying so many decimals.
But ckUSDC only has 6 decimals, so we should add 2 meaningless zeros to
display 8.

# Changes

1. Change `formatTokenE8s` to `formatTokenUlps` and make it generic in
the number of decimals.
2. Call `formatTokenUlps` from `formatTokenE8s` with specific values for
the number of displayed and detailed decimals.
3. Call `formatTokenUlps` instead of `formatTokenE8s` from
`formatTokenV2`.

# Tests

1. Added unit tests.
2. Looked manually that the transaction screen looks good.


# Todos

- [x] Add entry to changelog (if necessary).