Support fewer than 8 decimals in AmountInput (#4890)
# Motivation

When entering an amount we allow entering up to 8 decimals.
Even though ckETH supports 18 decimals, we decided to support 8.
But ckUSDC only supports 6 decimals so we don't want to allow entering
more than 6.
`Input` from gix-component already supports choosing the number of
decimals if you use `inputType = "currency"` instead of `"icp"` but then
the value is given as a string instead of a number.

# Changes

1. Allow setting `decimals` and `inputType = "currency"` on `Input` and
`InputWithError`.
2. Use `inputType = "currency"` in `AmountInput` and convert back to a
number.
3. Accept an optional `token` prop on `AmountInput` and pass `decimals`
based on the token.

# Tests

Added unit tests with ckETH, ckUSDC and without token.

# Todos

- [x] Add entry to changelog (if necessary).