Create restore instructions from snapshot install operation (#4895)
# Motivation

When installing an snsdemo snapshot we moving a bunch of directories.
When uninstalling the snapshot we move those directories in the opposite
direction in reverse order.
These moves are coded separately, which is fragile if we change what
needs to be moved for the install.
We want to change what's installed so existing identities (such as
`hsm`) can still be used while a snapshot is running.

In this PR we don't change what's installed but we generate the restore
instructions from what's installed so it's less likely to get out of
sync.

# Changes

1. When creating a directory, record a `rmdir` command for the restore.
2. When moving a file or directory, record a reverse `mv` command for
the restore.
3. When generating the restore script, output the recorded restore
commands in reverse order.
4. Use trap to create the restore script because the restore script can
no longer be created before the install is performed, but we want to
make sure the (partial) script is still output even if something goes
wrong with the install.

# Tests

Ran `scripts/dfx-snapshot-install` and inspected and ran the created
`restore.sh` script.
Also placed an `exit 1` after installing just the data directory and
checked that the partial restore script is correct.

# Todos

- [ ] Add entry to changelog (if necessary).
no necessary