Pass column specification to ResponsiveTable (#4919)
# Motivation

We want to make `ResponsiveTable` generic to reuse it for the neurons
table.
For this we need to customize which columns will be rendered for each
row.

# Changes

1. Add type `ResponsiveTableColumn` which specifies a title for a single
column and how to render cells in that column.
2. Pass a list of `ResponsiveTableColumn` from `TokensTable` to
`ResponsiveTable`.
3. In `ResponsiveTable` render column headers based on the column
specification.
4. In `ResponsiveTableRow` use components from the column specification
to render the cells.

# Tests

Existing tokens table tests pass.
Will add tests for responsive table when it's fully generic.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary