Support styling different numbers of columns in ResponsiveTable (#4951)
# Motivation

We want to reuse the `ResponsiveTable`, that was extracted from the
`TokensTable`, for the neurons table.
The neurons table has a different number of columns, which requires
different styling.

For desktop we give every column as much space as needed, except the
first column which will get all the remaining space.
We do this with a `grid-template-columns` style like `1fr max-content
max-content` where `1fr` assigns the remaining space to the first column
and we add additional `max-content` copies until we match the number of
columns.

For mobile each single row on desktop becomes a block of rows, like
this:
```
+------------+-----------+
| first-cell | last-cell |
| Label 0    |    cell-0 |
| Label 1    |    cell-1 |
| etc.                   |
+------------+-----------+
```
We accomplish this with a `grid-template-areas` style like `"first-cell
last-cell" "cell-0 cell-0" "cell-1 cell-1"`.

Both of these are assigned to CSS variables and then used in the
appropriate place.

Lastly, cells need to have their correct `grid-area` assigned as
`cell-${i}` and this is again done through CSS variables.
For desktop the grid area is reverted because we don't use the multi-row
style.

# Changes

1. Compute and assign CSS variables in
`frontend/src/lib/components/ui/ResponsiveTable.svelte`.
2. Use CSS variables to create the correct style.
3. Defined `getCellGridAreaName` in
`frontend/src/lib/utils/responsive-table.utils.ts` for consistent
`grid-area` naming.

# Tests

1. Added unit tests for the assigned styles.
2. Tested manually on the tokens table at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary