NNS1-3091: Make rowHref optional on ResponsiveTable (#4957)
# Motivation

When rendering table rows in a `{#each}` we need to assign a uniq DOM
key to each row so Svelte can keep track of rows for the purpose of
transitions.
For the tokens table we used the token URL for this because each row had
a unique URL anyway.
But for the neurons table, we need to be able to render rows for
spawning neurons, which do not have a details page and so do not have a
URL.
So we need another way to assign the DOM key without have a URL.

# Changes

1. Add a new `domKey` field to `ResponsiveTableRowData`.
2. Make `rowHref` optional.
3. Use `domKey` instead of `rowHref` in the `{#each}` in
`frontend/src/lib/components/ui/ResponsiveTable.svelte`.
4. Render rows which do not have a URL as `div` instead of `a`. (This
was also done before https://github.com/dfinity/nns-dapp/pull/4647)
5. Remove the `@include interaction.tappable;` style. Links already have
the correct style (pointer cursor) and non-links should not have this
style.
6. For the tokens table data, assign `domKey` everywhere to be identical
to `rowHref`. Something else could work as well but this is what we've
been using and it works.

# Tests

1. Assign `domKey` equal to `rowHref` in all the tokens table tests.
2. Add a unit test to check that `rowHref` can be absent and results in
a `<div>` row instead of `<a>`.
3. Manually tests by arbitrarily omitting some `rowHref`s in the tokens
table.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary